рефакторинг

+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserDbStorage.java
@@ -5,6 +5,7 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exceptions.FoundException;
+import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.users.User;
 import ru.yandex.practicum.filmorate.storage.LikesDao;
 
@@ -20,6 +21,7 @@ public class UserDbStorage implements UserStorage {
 
     private final JdbcTemplate jdbcTemplate;
     private final LikesDao likesDao;
+
     @Autowired
     public UserDbStorage(JdbcTemplate jdbcTemplate, LikesDao likesDao) {
         this.jdbcTemplate = jdbcTemplate;
@@ -28,7 +30,9 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public User create(User user) {
-
+        if (emailValidation(user.getEmail())) {
+            throw new ValidationException("This email is used");
+        }
         String sqlQuery = "insert into users (email, login, name, birthday) " +
                 "values (?, ?, ?, ?)";
         jdbcTemplate.update(sqlQuery,
@@ -51,7 +55,7 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public User update(User user) {
-        idValidation(user.getId());
+        userIdValidation(user.getId());
         String sqlQuery = "update users set name = ?, login = ?, birthday = ?, email = ?" +
                 "where user_id = ?";
         jdbcTemplate.update(sqlQuery,
@@ -65,52 +69,58 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public User getById(Long id) {
-        idValidation(id);
+        userIdValidation(id);
         String sqlQuery = "select * from users where user_id = ?";
         return jdbcTemplate.queryForObject(sqlQuery, (rs, rowNum) -> makeUser(rs), id);
     }
 
     @Override
-    public void saveLike(long filmId, long userId) {
-        likesDao.saveLike(filmId, userId);
-    }
-
-    @Override
-    public void deleteLike(long filmId, long userId) {
-        likesDao.deleteLike(filmId, userId);
-    }
-
-    @Override
     public void delete(String email) {
+        if (!emailValidation(email)) {
+            throw new FoundException("User with this email is not registered");
+        }
         String sqlQuery = "delete from users where email = ?";
         jdbcTemplate.update(sqlQuery, email);
     }
+
     @Override
-    public void saveFriend(long userId, long friendId){
+    public void saveFriend(long userId, long friendId) {
+        userIdValidation(userId);
+        userIdValidation(friendId);
         String sqlQuery = "merge into friends (user_id, friend, friendship_status) key (user_id, friend)" +
                 "values (?,?,?)";
         jdbcTemplate.update(sqlQuery, userId, friendId, 1);
     }
+
     @Override
-    public void saveSubscriber(long userId, long subscriberId){
+    public void saveSubscriber(long userId, long subscriberId) {
+        userIdValidation(userId);
+        userIdValidation(subscriberId);
         String sqlQuery = "merge into friends (user_id, friend, friendship_status) key (user_id, friend)" +
                 "values (?,?,?)";
         jdbcTemplate.update(sqlQuery, userId, subscriberId, 2);
     }
+
     @Override
     public void deleteFriend(long userId, long subscriberId) {
+        userIdValidation(userId);
+        userIdValidation(subscriberId);
         String sqlQuery = "delete from friends where user_id = ? and friend = ?";
         jdbcTemplate.update(sqlQuery, userId, subscriberId);
     }
+
     @Override
     public List<Long> getUserFriends(long id) {
+        userIdValidation(id);
         String sqlQuery = "select friend from friends where user_id = ? and friendship_status = 1";
         return jdbcTemplate.queryForList(sqlQuery, Integer.class, id).stream()
                 .map(Long::valueOf)
                 .collect(Collectors.toList());
     }
+
     @Override
     public List<Long> getUserSubscribers(long id) {
+        userIdValidation(id);
         String sqlQuery = "select friend from friends where user_id = ? and friendship_status = 2";
         return jdbcTemplate.queryForList(sqlQuery, Integer.class, id).stream()
                 .map(Long::valueOf)
@@ -130,12 +140,19 @@ public class UserDbStorage implements UserStorage {
         return user;
     }
 
-    private void idValidation(long id) {
+    private void userIdValidation(long id) {
         List<Integer> userIds;
         String sqlQuery = "select user_id from users";
         userIds = jdbcTemplate.queryForList(sqlQuery, Integer.class);
         if (!userIds.contains((int) id)) {
             throw new FoundException("Unknown user");
         }
+    }
+
+    private boolean emailValidation(String email) {
+        List<String> emails;
+        String sqlQuery = "select email from users";
+        emails = jdbcTemplate.queryForList(sqlQuery, String.class);
+        return emails.contains(email);
     }
 }
